requirements.txt
streamlit
langchain
sentence-transformers
tiktoken
faiss-cpu
beautifulsoup4
requests
#deepgram-sdk
langchain_community
moviepy
deepgram-sdk==2.12.0
langchain_groq
#Deepgram Api key
#551e5521166984fcbd8207d439416865d4f2f457